backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
paho-mqtt>=2.1.0,<3.0.0
aiomqtt==2.5.0
redis==5.0.1
numpy==2.4.1
scipy==1.17.0
pandas==2.2.3
influxdb-client==1.39.0
python-dotenv==1.0.0
PyYAML==6.0.1
websockets==12.0
python-socketio==5.11.2
python-engineio==4.8.0
python-multipart==0.0.6
httpx==0.25.2